I found the problem with extracting multiple distribution sets. Mac OS X's tar (in 10.4.8 anyway) is GNU tar, whereas FreeBSD has used BSD tar as the default /usr/bin/tar since FreeBSD 5. FreeBSD 1-4 I believe used GNU tar for some reason. GNU tar handles the --unlink option in a different way than BSD tar. With GNU tar, my previous attempts to extract multiple dists would lead to tar trying to first remove directories that already existed and contained files, which gave errors and made the extraction fail.
Here's a tar wrapper that gets rid of the --unlink option on non-FreeBSD
systems, and hopefully will also make it possible to run these tests on Linux or
other operating systems that use a totally different tar and/or don't put it in
/usr/bin. I haven't actually tested the tests (heh) on Linux or Solaris or
anything like that yet.

Note that FreeBSD systems which use GNU tar will still have problems using tar
to extract these files, and the extraction test will fail. I think it is safe to
let farbot be incompatible (it might have already been anyway) with FreeBSD 4
and lower, though.